Reshuffled Folder Structure, added Main Menu
New folder structure. New main menu scene with basic title, play button, high score text. Added score text to gameplay scene.

Also made a SceneManager object with a SceneSwitcher script. It has a "SwitchToScene" function the play button uses to switch to Gameplay scene, but this may not be the optimal way to do it...